#4797 Removed redundant ValueExpressionAnalyzer
Moved some tests in order to test result
