--- test/javaee8/validateWholeBean/src/main/java/com/sun/faces/test/javaee8/validateWholeBean/PasswordValidator.java
@@ -1,37 +1,0 @@
-/*
- * Copyright (c) 1997, 2018 Oracle and/or its affiliates. All rights reserved.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v. 2.0, which is available at
- * http://www.eclipse.org/legal/epl-2.0.
- *
- * This Source Code may also be made available under the following Secondary
- * Licenses when the conditions for such availability set forth in the
- * Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
- * version 2 with the GNU Classpath Exception, which is available at
- * https://www.gnu.org/software/classpath/license.html.
- *
- * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
- */
-
-package com.sun.faces.test.javaee8.validateWholeBean;
-
-import java.util.logging.Logger;
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
-
-public class PasswordValidator implements ConstraintValidator<Password, BackingBean> {
-    
-    @Override
-    public void initialize(Password constraintAnnotation) {
-        // NOOP
-    }
-
-    @Override
-    public boolean isValid(BackingBean value, ConstraintValidatorContext context) {
-        boolean result;
-        result = value.getPassword1().equals(value.getPassword2());
-        return result;
-    }
-
-}
