--- test/javaee8/validateWholeBean/src/main/java/com/sun/faces/test/javaee8/validateWholeBean/UserPassword.java
@@ -1,38 +1,0 @@
-/*
- * Copyright (c) 1997, 2018 Oracle and/or its affiliates. All rights reserved.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v. 2.0, which is available at
- * http://www.eclipse.org/legal/epl-2.0.
- *
- * This Source Code may also be made available under the following Secondary
- * Licenses when the conditions for such availability set forth in the
- * Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
- * version 2 with the GNU Classpath Exception, which is available at
- * https://www.gnu.org/software/classpath/license.html.
- *
- * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
- */
-
-package com.sun.faces.test.javaee8.validateWholeBean;
-
-import static java.lang.annotation.ElementType.TYPE;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-import javax.validation.Constraint;
-import javax.validation.Payload;
-
-@Constraint(validatedBy = {UserPasswordValidator.class})
-@Documented
-@Target(TYPE)
-@Retention(RUNTIME)
-public @interface UserPassword {
-
-    String message() default "Password fields must match";
-    Class<?>[] groups() default {};
-    Class<? extends Payload>[] payload() default {};
-}
