--- test/javaee8/validateWholeBean/src/main/java/com/sun/faces/test/javaee8/validateWholeBean/UserPasswordValidator.java
@@ -1,36 +1,0 @@
-/*
- * Copyright (c) 1997, 2018 Oracle and/or its affiliates. All rights reserved.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v. 2.0, which is available at
- * http://www.eclipse.org/legal/epl-2.0.
- *
- * This Source Code may also be made available under the following Secondary
- * Licenses when the conditions for such availability set forth in the
- * Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
- * version 2 with the GNU Classpath Exception, which is available at
- * https://www.gnu.org/software/classpath/license.html.
- *
- * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
- */
-
-package com.sun.faces.test.javaee8.validateWholeBean;
-
-import java.util.Objects;
-
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
-
-public class UserPasswordValidator implements ConstraintValidator<UserPassword, User> {
-    
-    @Override
-    public void initialize(UserPassword constraintAnnotation) {
-        // NOOP
-    }
-
-    @Override
-    public boolean isValid(User user, ConstraintValidatorContext context) {
-        return Objects.equals(user.getPassword1(), user.getPassword2());
-    }
-
-}
