--- test/javaee8/validateWholeBean/src/test/java/com/sun/faces/test/javaee8/validateWholeBean/Issue4313IT.java
@@ -1,71 +1,0 @@
-/*
- * Copyright (c) 1997, 2018 Oracle and/or its affiliates. All rights reserved.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v. 2.0, which is available at
- * http://www.eclipse.org/legal/epl-2.0.
- *
- * This Source Code may also be made available under the following Secondary
- * Licenses when the conditions for such availability set forth in the
- * Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
- * version 2 with the GNU Classpath Exception, which is available at
- * https://www.gnu.org/software/classpath/license.html.
- *
- * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
- */
-
-package com.sun.faces.test.javaee8.validateWholeBean;
-
-import static org.junit.Assert.assertEquals;
-
-import org.junit.After;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
-
-import com.gargoylesoftware.htmlunit.WebClient;
-import com.gargoylesoftware.htmlunit.html.HtmlPage;
-import com.gargoylesoftware.htmlunit.html.HtmlPasswordInput;
-import com.gargoylesoftware.htmlunit.html.HtmlSubmitInput;
-
-public class Issue4313IT {
-
-    private String webUrl;
-    private WebClient webClient;
-
-    @Before
-    public void setUp() {
-        webUrl = System.getProperty("integration.url");
-        webClient = new WebClient();
-    }
-
-    @After
-    public void tearDown() {
-        webClient.closeAllWindows();
-    }
-
-    @BeforeClass
-    public static void setUpClass() {
-    }
-
-    @AfterClass
-    public static void tearDownClass() {
-    }
-
-    @Test
-    public void testValidateWholeBean() throws Exception {        
-        HtmlPage page = webClient.getPage(webUrl + "faces/Issue4313.xhtml");
-
-        HtmlPasswordInput password1 = (HtmlPasswordInput) page.getHtmlElementById("form:password1");
-        password1.setValueAttribute("mike@!2016GO");
-
-        HtmlPasswordInput password2 = (HtmlPasswordInput) page.getHtmlElementById("form:password2");
-        password2.setValueAttribute("mike@!2015G0");
-
-        HtmlSubmitInput submit = (HtmlSubmitInput) page.getHtmlElementById("form:submit");
-        HtmlPage page1 = submit.click();
-
-        assertEquals("Password fields must match", page1.getElementById("form:err").getElementsByTagName("li").get(0).asText());
-    }
-}
