+++ test2/faces22/cdiBeanValidator/src/main/webapp/WEB-INF/web.xml
