+++ test2/faces23/validateWholeBean/pom.xml
@@ -17,28 +17,22 @@
     SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
 
 -->
+
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>org.glassfish.mojarra.test2</groupId>
+        <groupId>org.glassfish.mojarra.faces23</groupId>
         <artifactId>pom</artifactId>
         <version>4.0.0-SNAPSHOT</version>
     </parent>
 
-    <groupId>org.glassfish.mojarra.faces23</groupId>
-    <artifactId>pom</artifactId>
-    <packaging>pom</packaging>
+    <artifactId>validateWholeBean</artifactId>
+    <packaging>war</packaging>
+    
+    <name>Mojarra ${project.version} - Test - Faces 2.3 - validateWholeBean</name>
 
-    <name>Mojarra ${project.version} - Test - Faces 2.3</name>
-
-    <modules>
-        <module>el</module>
-        <module>exactMapping</module>
-        <module>faceletCacheFactory</module>
-        <module>facelets</module>
-        <module>getViews</module>
-        <module>systemEvent</module>
-        <module>validateWholeBean</module>
-  </modules>
+    <build>
+        <finalName>test-faces23-systemEvent</finalName>
+    </build>
 </project>
