+++ test2/faces23/validateWholeBean/src/main/webapp/Issue4313.xhtml
