+++ test2/faces23/validateWholeBean/src/main/webapp/WEB-INF/beans.xml
