+++ test2/faces23/validateWholeBean/src/main/webapp/index.xhtml
